Call overridden version of `to_liquid` and remove duplicate code
Make sure to call "overidden" method which keeps us nice and compatible
so we don't need to keep up with whatever Jekyll usually does, and make
sure to include it in our overriden method.

I have had it with these monkey-patching methods in this Monday-to-
Friday language!

Add method `liquid_date_attributes`.

Cleanup the Post and Page overrides to be more similar.